Add sma up and down strategy

+++ pyalgotrade/broker/__init__.py
@@ -40,7 +40,7 @@ class InstrumentTraits(object):
 
 class IntegerTraits(InstrumentTraits):
     def roundQuantity(self, quantity):
-        return int(quantity)
+        return float(quantity)
 
 
 ######################################################################
